OSX fixes.

+++ implementations/meth1/stratergies/meth1/meth1_memory.hh
@@ -5,7 +5,7 @@
 #include <cstdint>
 
 /* Get a count of how many disks this machine has */
-size_t num_of_disks( void );
+uint64_t num_of_disks( void );
 
 /* figure out max chunk size available */
 uint64_t calc_record_space( void );
